LPS-114698 Rename to match: git grep "implements ContextProvider"
